fixed variable type to long long

+++ src/WAVTrack.cpp
@@ -15,7 +15,7 @@ void WAVTrack::load() {
     // NOTE: Use exactly 2 spaces before the arrow (→) character
     std::cout << "[WAVTrack::load] Loading WAV: \"" << title << "\" at " 
           << sample_rate << "Hz/" << bit_depth <<"bit (uncompressed)...\n";
-    long size = duration_seconds * sample_rate * (bit_depth / 8) * 2;
+    long long size = duration_seconds * sample_rate * (bit_depth / 8) * 2;
     std::cout <<"  -> Estimated file size: \"" << size << "\" bytes \n";
     std::cout <<"  -> Fast loading due to uncompressed format. \n";
 }
